+++ CHANGELOG.md
@@ -6,11 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * [Unreleased](#unreleased)
+* [0.4.0](#040---2026-01-24)
 * [0.3.0](#030---2026-01-21)
 * [0.2.0](#020---2026-01-20)
 * [0.1.0](#010---2026-01-18)
 
 ## [Unreleased]
+
+### Added
+
+*   Added support for 2d and 2dsphere indexes in `com.github.starnowski.jamolingo.perf.ExplainAnalyzeResultFactory` ([#18](https://github.com/starnowski/jamolingo/issues/18))
+*   Added support for range filtering in `com.github.starnowski.jamolingo.perf.ExplainAnalyzeResultFactory` ([#18](https://github.com/starnowski/jamolingo/issues/18))
+
+
+## [0.4.0] - 2026-01-24
 
 ### Added